SEEK: Following changes have been made for SEEK
1. Support getAtr api.
   Implement getAtr for SEEK on Android.

2. Send correct response for iccTransmitApduLogicalChannel,
   iccTransmitApduBasicChannel & iccExchangeSimIO requests.

   When these requests fail, PhoneInterfaceManager sets its
   response to a default value ( IccIoResult with sw1 = 0x6f,
   sw2 = 0  and payload as null).

   Do not modify the response of the requests; Send the
   response received from RIL to the client.

3. Add Multi sim support.

   Msim support for following apis has been added:
   1. iccOpenLogicalChannel
   2. iccCloseLogicalChannel
   3. iccTransmitApduLogicalChannel
   4. iccTransmitApduBasicChannel
   5. iccExchangeSimIO
   6. getATR

CRs-Fixed: 718469

4. SEEK: Send correct responses for apdu transmit and simIO requests.

Correct the responses sent for IccTransmitApduLogicalChannel,
iccTransmitApduBasicChannel & iccExchangeSimIO requests:

When these requests fail with some response data from RIL,
send the response received from RIL to client.

But when the requests fail without any response data from RIL,
set the response to a default value ( IccIoResult with sw1 = 0x6f,
sw2 = 0  and payload as null). Without this, the response for the client
is set to null and PhoneInterfaceManager enters an infinite loop while
processing the response.

CRs-Fixed: 829284
Change-Id: I9494e17395a834b9a7c997c1afb5ef4b31b28e76

5. Msim SEEK: Make sure the apis use the subId correctly.

Retireve the uicc card object and check for enforced permissions based
on the subId provided by the client in the apis.

Change-Id: Idedb0a62a950e923a7a3fde5d07492f0c7a125a3
CRs-Fixed: 840826

Change-Id: I794e614e324c414b08c78fbdd601414b52b98b77